Refactor: Extract BashScript type
This encapsulates a user-provided commit script, and how to execute it with bash. This lets us pass more information around (e.g. line number, for better error messages), and also provides a more obvious extension point for supporting other script languages in future.
